Fixed two bugs with mercy-illegal
The first bug is related to Config.GrowthObjects: In the GetStageFromPlant function you return stages between 1 and 5, but the array had values between 0 and 4. At stage 5 it could not find the correct model to use.

The second bug is related to the GetStageFromPlant function. Because of the randomness in the plant growth (a value between 3 and 6) the number could be higher than 100. The function would then not return a stage at all.